🔨 Refactor : MyPage 코드 스플릿
